refactor(canary): Organize text generation prompts
- Move text generation prompts from individual component file to a
  centralized `lib/prompts` directory for improved organization and
  clarity. This refactoring facilitates easier access and maintenance of
  prompt-related functions.

- Update import statements in affected files to reference the new
  location of the prompts.

No breaking changes introduced with this update.
